+++ tests/testthat/test-period.R
@@ -33,7 +33,7 @@ test_that("diff()", {
   expect_identical(interval_pull(diff(yq) + c(3, 6)), new_interval(quarter = 3))
 })
 
-test_that("Arit", {
+test_that("Arithmetic", {
   expect_equal(format(+yw), format(yw))
   expect_equal(format(yw + 1), c("1970 W02", "1970 W03", "1970 W04"))
   expect_equal(format(1 + yw), format(yw + 1))
@@ -129,4 +129,22 @@ test_that("yearquarter() for characters #107", {
 
 test_that("yearquarter.character() underlying dates #129", {
   expect_equal(as.Date(yearquarter("2017 Q1")), as.Date("2017-01-01"))
+})
+
+test_that("yearweek() for characters", {
+  expect_error(yearweek("2013 We 3"), "cannot be expressed as Date type")
+  expect_error(yearweek("Wee 5 2015"), "cannot be expressed as Date type")
+  expect_error(yearweek("W54 2015"), "can't be greater than 53.")
+  expect_error(yearweek(c("2015 W53", "2016 W53", "2017 W53")), "can't be 53 weeks.")
+  expect_error(yearweek("W2015"), "unambiguous")
+  expect_error(yearweek(c("W2015", "W2 2015")), "unambiguous")
+  expect_identical(
+    yearweek(c("2013 W3", "2013 Wk 3", "Week 3 2013")),
+    rep(yearweek("2013 W03"), 3)
+  )
+})
+
+test_that("yearweek.character() underlying dates", {
+  expect_equal(as.Date(yearweek("1970 W01")), as.Date("1969-12-29"))
+  expect_equal(as.Date(yearweek("2019 W12")), as.Date("2019-03-18"))
 })